+++ NEWS.md
@@ -5,6 +5,8 @@
   extractXi.
 * hullplot uses now filled polygons as the default.
 * hullplot now used PCA if the data has more than 2 dimensions.
+* kNN and frNN are now objects with a print function.
+
 
 # Changes in version 0.9-8 (2016-08-05)